IOS-4990: Improve project file structure
